Fix cleaning code

See merge request hpc-ai_21/cpp-life-of-boids!52

+++ src/lib/myMath/utils.cpp
@@ -1,9 +1,3 @@
-//#include "utils.hpp"
-////#define _USE_MATH_DEFINES
-////#include <cmath>
-//#include "Vec2.hpp"
-//#include <vector>
-//#include <random>
 #include "Vec2.hpp"
 #include <vector>
 #include <random>
@@ -59,14 +53,14 @@ int researchX(const float& target, const std::vector<std::tuple<float, float, in
   if (std::get<0>(v[0]) >= target){
     return 0;
   }
-  else if (std::get<0>(v[size - 1]) <= target) {
+  else if (std::get<0>(v[(int)(size - 1.0)]) <= target) {
     return size - 1;
   }
   else{
     int d = 4;
     int a;
     int middle = (int) size / 2;
-    while (!(std::get<0>(v[middle - 1]) < target  && target <= std::get<0>(v[middle]))){
+    while (!(std::get<0>(v[(int)(middle - 1.0)]) < target  && target <= std::get<0>(v[middle]))){
       a = (int)size / d;
       if (target < std::get<0>(v[middle])) {
         middle -= (a!=0) ? a : 1;
@@ -85,14 +79,14 @@ int researchY(const float& target, const std::vector<std::tuple<float, float, in
   if (std::get<1>(v[0]) > target){
     return 0;
   }
-  else if (std::get<1>(v[size - 1]) <= target) {
+  else if (std::get<1>(v[(int)(size - 1.0)]) <= target) {
     return size - 1;
   }
   else{
     int d = 4;
     int a;
     int middle = (int) size / 2;
-    while (!(std::get<1>(v[middle - 1]) <= target  && target <= std::get<1>(v[middle]))){
+    while (!(std::get<1>(v[(int)(middle - 1.0)]) <= target  && target <= std::get<1>(v[middle]))){
       a = (int) size / d;
       if (target < std::get<1>(v[middle])) {
         middle -= (a!=0) ? a : 1;
